repos/at-chaos-tests/requirements.txt
# Chaos Testing Framework Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
structlog==23.2.0
prometheus-client==0.19.0
nats-py==2.6.0
aiohttp==3.9.1
psutil==5.9.6
numpy==1.24.3
pyyaml==6.0.1
docker==6.1.3
asyncio-mqtt==0.16.1